requirements.txt
Django==1.7.1
django-mptt==0.6.1
psycopg2==2.5.4
wsgiref==0.1.2
